move `SmalltalkTestStatus` from one enum module to another
