+++ samanthas_telegram_bot/api_clients/auxil/enums.py
@@ -15,3 +15,11 @@ class LoggingLevel(str, Enum):
     ERROR = "error"
     CRITICAL = "critical"
     EXCEPTION = "exception"
+
+
+class SmalltalkTestStatus(str, Enum):
+    """Enumeration of possible statuses of a SmallTalk oral test, returned by its API."""
+
+    NOT_STARTED_OR_IN_PROGRESS = "sent"  # meaning "link to interview was sent to user"
+    RESULTS_NOT_READY = "processing"
+    RESULTS_READY = "completed"
