+++ samanthas_telegram_bot/data_structures/models.py
@@ -6,8 +6,9 @@ from dataclasses import dataclass
 from typing import Optional, TypedDict
 
 from samanthas_telegram_bot.api_clients.auxil.constants import DataDict
+from samanthas_telegram_bot.api_clients.auxil.enums import SmalltalkTestStatus
 from samanthas_telegram_bot.data_structures.constants import WEEKDAYS
-from samanthas_telegram_bot.data_structures.enums import AgeRangeType, SmalltalkTestStatus
+from samanthas_telegram_bot.data_structures.enums import AgeRangeType
 
 
 @dataclass
